feat: Add support for V&C Dungeons
V&C Dungeons are new types of dungeons added in 6.25.

Square added a new category flag, 128, in order to identify them. Looking at
the listings send to the server, there seems to be no difference between
Variant, Critereon and Critereon Savage dungeons in terms of categorization
flags.

Translations were copied from the patch notes and later verified in-game.

V&C Dungeons are now shown in the frontend. The category exists and
filtering works as expected.